agents/requirements.txt
uagents
google-genai
python-dotenv
